Auth routes split
